feat: Item routing and user service functionality
- Added item routing to `index.js` for handling item-related API endpoints.
- Updated user service methods in `user_router.js` to return user objects instead of categories.
- Modified the average rating validation in `item_schema.js` to use a number type instead of float.

+++ server/schemas/item_schema.js
@@ -7,7 +7,7 @@ const description = Joi.string().max(500);
 const coverUrl = Joi.string().max(255);
 const releaseDate = Joi.date();
 const externalId = Joi.number().integer();
-const avgRating = Joi.number().float();
+const avgRating = Joi.number();
 const vibeTags = Joi.array().items(Joi.string());
 
 const createItemSchema = Joi.object({
